Revert "update seperation of concerns and improve default exclusion"
This reverts commit ad65a191e7e873d8ddf09f8f119e7efa0b8626bb.
